Namespace the container resolver into Linux
FreeBSD has containers (jails) but already handled through the 'virutal'
resolver, so for now we just consider Linux being able to handle
containers as the code aleady depend on Linux-specific resolvers.
