+++ lib/facter/framework/core/file_loader.rb
@@ -52,7 +52,6 @@ require_relative '../../resolvers/aio_agent_version'
 require_relative '../../resolvers/augeas'
 require_relative '../../resolvers/az'
 require_relative '../../resolvers/base_resolver'
-require_relative '../../resolvers/containers'
 require_relative '../../resolvers/debian_version'
 require_relative '../../resolvers/dmi_decode'
 require_relative '../../resolvers/ec2'
@@ -503,6 +502,7 @@ os_hierarchy.each do |os|
     require_relative '../../facts/linux/virtual'
     require_relative '../../facts/linux/xen'
 
+    require_relative '../../resolvers/linux/containers'
     require_relative '../../resolvers/linux/disks'
     require_relative '../../resolvers/linux/dmi'
     require_relative '../../resolvers/linux/docker_uptime'
