+++ spec/facter/facts/linux/hypervisors/docker_spec.rb
@@ -5,7 +5,7 @@ describe Facts::Linux::Hypervisors::Docker do
     subject(:fact) { Facts::Linux::Hypervisors::Docker.new }
 
     before do
-      allow(Facter::Resolvers::Containers).to \
+      allow(Facter::Resolvers::Linux::Containers).to \
         receive(:resolve).with(:hypervisor).and_return(hv)
     end
